pressure-alarm-test-harness/src/main/Alarm.java
package main;

public class Alarm
{
    private final double lowPressureThreshold = 17;
    private final double highPressureThreshold = 21;

    private PsiSensor sensor = new Sensor();

    private boolean alarmOn = false;
    
    public Alarm(PsiSensor psiSensor) {
    	this.sensor = psiSensor;
    }

    public void check()
    {
        double psiPressureValue = sensor.popNextPressurePsiValue();

        if (psiPressureValue < lowPressureThreshold || highPressureThreshold < psiPressureValue)
        {
            alarmOn = true;
        }
    }

    public boolean isAlarmOn()
    {
        return alarmOn;
    }
}
